Work on rt combat
added the text and the ui bits for it so turn based is pretty much complete.

+++ Assets/gameScript/Combat/IKillable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IKillable
 {
-    public abstract void TakeDamage(int amount);
+    public abstract void TakeDamage(float amount);
     public abstract bool IsDead();
     public abstract void Die();
 }
